Replace MIN_SLICE_LEN with ALIGNMENT in the device memory (#503)
This main purpose of this change to support arbitrary small slices of the "device" memory in `FastCpuComputeLayer`.

The main idea is to represent `FSlice`/`FSliceMut` as an enum with two options:
 - Borrowed slice, when the number of elements is a multiple of `P::WIDTH`
 - Owned storage with a single packed element for the case when the length is strictly less than `P::WIDTH`.

We still can't represent slices bigger than `P::WIDTH` which aren't aligned by the `P::WIDTH` border but that seems to be sufficient for our needs, e.g. we are always able to split pow of two length slices by half.

The changes introduced by this PR:
 - replace `ComputeMemory::MIN_SLICE_LEN` with `ComputeMemory::ALIGNMENT`
 - relax conditions for all the slice spliting method, now the slices with the non-aligned length are allowed when their length is strictly less than the alignment
 - change the implementation of `FSlice` and `FSliceMut` in the `binius_fast_compute` crate.

The HW and `CPULayer` implementations of the memory do not need to change anything apart from applying the renaming.