[fri] Choose FRI domains to be folding compatible (#549)
This PR changes the FRI parameter selection logic to choose Reed-Solomon evaluation domains that are NTT-compatible, regardless of the trace size.

Previously, the evaluation subspace depended on the trace size. Now, we choose the FRI domains to be the push-forwards of a stable domain (the T_5 basis) under the q maps. This will assist with batch FRI arguments with pre-committed tables.