[prodcheck] Prodcheck module with layer computation (#672)
Introduce new module for the ComputeLayer-accelerated prodcheck protocol. This first commit contains code and a test for computing the circuit layer evaluations.

I'm implementing this as a new module instead of a modification to `gkr_gpa`. This is because I will reimplement the protocol verifier in a following PR, so this is a full-scale rewrite of the protocol.