[core] push TableSizeSpec down to core (#704)
Pushes table size spec down to binius_core. Now, the table sizes are also
checked as part of `prove`. The check in m3 stays for now, but it is going away
soon. The check in `verify` will come at a later time, for that we would need to
put the sizes onto transcript.

Similar to PR #581.