table-aware constraint system (#721)
This is the last PR in the series that implements
[CRY-363: Table-Aware Constraint System](https://linear.app/irreducible/issue/CRY-363/table-aware-constraint-system)

In this PR we introduce a structure called SymbolicMultilinearOracleSet. This
structure is very similar to MultilinearOracleSet with the exception that it
does not have any information about the sizes (n_vars) of the declared
multilinears.

This symbolic representation is what defines the ConstraintSystem, instead of
the concrete one. That, and other changes in this PR, make constraint system
serialization independent of the sizes of tables.

This PR avoids making large changes to the proving path and to achieve that I
decided to take the strategy in which the multilinears for zero-sized tables
are not created at all.